backend/requirements.txt
ansible==7.1.0
ansible-core==2.14.1
anyio==3.6.2
asgiref==3.6.0
attrs==22.2.0
autobahn==23.1.1
Automat==22.10.0
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.0.1
ciscoconfparse==1.7.9
click==8.1.3
colorama==0.4.6
constantly==15.1.0
coverage==7.1.0
cryptography==39.0.0
daphne==4.0.0
deprecat==2.1.1
Django==4.1.5
django-bootstrap5==22.2
django-cors-headers==3.13.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.3.0
et-xmlfile==1.1.0
exceptiongroup==1.1.0
future==0.18.3
greenlet==2.0.2
h11==0.12.0
httpcore==0.13.7
httpx==0.18.2
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
iniconfig==2.0.0
Jinja2==3.1.2
junos-eznc==2.6.6
loguru==0.6.0
lxml==4.9.2
MarkupSafe==2.1.2
mypy-extensions==0.4.3
napalm==4.0.0
ncclient==0.6.13
netaddr==0.8.0
netmiko==4.1.2
netutils==1.4.0
nornir==3.3.0
nornir-cli==1.1.1
nornir-f5==0.5.1
nornir-http==0.1.1
nornir-jinja2==0.2.0
nornir-napalm==0.3.0
nornir-netbox==0.3.0
nornir-netconf==1.0.1
nornir-netmiko==0.2.0
nornir-paramiko==0.2.0
nornir-pyez==0.2.4
nornir-pyxl==1.0.1
nornir-routeros==0.4.0
nornir-scrapli==2022.7.30
nornir-utils==0.2.0
ntc-templates==3.2.0
openpyxl==3.0.10
packaging==20.9
paramiko==2.12.0
passlib==1.7.4
pluggy==1.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyeapi==0.8.4
PyJWT==2.6.0
PyNaCl==1.5.0
pyOpenSSL==23.0.0
pyparsing==3.0.9
pyserial==3.5
pytest==7.2.1
pytest-cov==3.0.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
requests-toolbelt==0.9.1
resolvelib==0.8.1
rfc3986==1.5.0
RouterOS-api==0.17.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scp==0.13.6
scrapli==2022.7.30
scrapli-cfg==2022.7.30
scrapli-community==2022.7.30
scrapli-netconf==2022.7.30
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.0
sqlparse==0.4.3
tenacity==8.1.0
textfsm==1.1.2
toml==0.10.2
tomli==2.0.1
tqdm==4.64.1
transitions==0.9.0
ttp==0.9.2
ttp-templates==0.3.2
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.4.0
urllib3==1.26.14
wrapt==1.14.1
xmltodict==0.12.0
yamlordereddictloader==0.4.0
zope.interface==5.5.2
